Pipelines | Add nature places, use caching on record level, rework fetching
