+++ pipelines/places/greenwalking/pipeline/places/fields.py
@@ -1,9 +1,11 @@
 ARTICLES = "articles"
 ARTIST = "artist"
+CLAIMS = "claims"
 GEOPOINT = "geopoint"
 COPYRIGHTED = "copyrighted"
 DESCRIPTION_URL = "descriptionUrl"
 GEOHASH = "geohash"
+IMAGE = "image"
 LATITUDE = "latitude"
 LICENSE_SHORT_NAME = "licenseShortName"
 LICENSE_URL = "licenseUrl"
