+++ pipelines/places/greenwalking/pipeline/places/wikidata/__init__.py
@@ -1,3 +1,2 @@
-from .fetch import Fetch
 from .query import Query
-from .process import ProcessDoFn as Process
+from .transform import Transform
